+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # BuildAssist
 
+## [1.1.0] - 2020-04-02
+
+### Added
+- Added items that can be set in WebGL.
+- Allow optional asset bundle builds to be selected
+- Add BuildAssistEvent and its sample
+- Allows selection of StreamingAssets copy after asset bundle build
+
+### Changed
+- Moved the platform selection to the project settings
+- Styles has been refactored.
+
+### Fixed
+- SymbolSettings linkage was not working, fixed.
+
 ## [1.0.1] - 2020-03-23
 
 ### Added